ui/requirements-ui.txt
streamlit==1.37.0
requests==2.32.0
pandas==2.3.0
plotly==5.23.0
streamlit-extras==0.7.5
streamlit-option-menu==0.3.6
python-dotenv==1.0.1